+++ src/Resources/contao/dca/tl_content.php
@@ -14,183 +14,116 @@
 /**
  * Palettes
  */
-$GLOBALS['TL_DCA']['tl_content']['palettes']['__selector__'][] = 'chesstable_lightbox';
-$GLOBALS['TL_DCA']['tl_content']['palettes']['chesstable'] = '{type_legend},type,headline;{chesstable_legend_csv},chesstable_csv;{chesstable_legend_aufab},chesstable_aufsteiger,chesstable_absteiger,chesstable_markieren;{chesstable_legend_lightbox},chesstable_lightbox;{chesstable_legend_optionen},chesstable_namendrehen,chesstable_flaggen,chesstable_date,chesstable_ende,chesstable_note;{protected_legend:hide},protected;{expert_legend:hide},guest,cssID,space;{invisible_legend:hide},invisible,start,stop';
-$GLOBALS['TL_DCA']['tl_content']['subpalettes']['chesstable_lightbox'] = 'chesstable_linktext,chesstable_hinweis';
+$GLOBALS['TL_DCA']['tl_content']['palettes']['chessresults'] = '{type_legend},type,headline;{chessresults_legend},chessresults_id,chessresults_type,chessresults_mode,chessresults_colsView,chessresults_round,chessresults_cache,chessresults_live;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID;{invisible_legend:hide},invisible,start,stop';
 
 /**
  * Fields
  */
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_csv'] = array
+$GLOBALS['TL_DCA']['tl_content']['fields']['chessresults_id'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_csv'],
+	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chessresults_id'],
+	'inputType'               => 'text',
+	'eval'                    => array
+	(
+		'mandatory'           => true,
+		'tl_class'            => 'w50',
+		'maxlength'           => 8,
+	),
+	'sql'                     => "int(8) unsigned NOT NULL default 0"
+);
+
+$GLOBALS['TL_DCA']['tl_content']['fields']['chessresults_mode'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chessresults_mode'],
 	'exclude'                 => true,
-	'search'                  => true,
-	'inputType'               => 'textarea',
+	'filter'                  => true,
+	'inputType'               => 'select',
+	'options'                 => &$GLOBALS['TL_LANG']['tl_content']['chessresults_mode_array'],
 	'eval'                    => array
 	(
-		'allowHtml'           => true,
-		'class'               => 'monospace',
-		'rows'                => 30,
-		'rte'                 => 'ace',
-		'helpwizard'          => true
+		'mandatory'           => true,
+		'includeBlankOption'  => true,
+		'chosen'              => true,
+		'tl_class'            => 'long clr'
 	),
-	'explanation'             => 'chesstable_csv',
-	'sql'                     => "mediumtext NULL",
+	'sql'                     => "int(2) unsigned NOT NULL default 0"
 );
 
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_file'] = array
+$GLOBALS['TL_DCA']['tl_content']['fields']['chessresults_round'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_file'],
-	'inputType'               => 'fileTree',
-	'eval'                    => array
-	(
-		'files'               => true,
-		'filesOnly'           => true,
-		'fieldType'           =>'radio',
-		'extensions'          => 'csv,txt',
-		'maxlength'           => 255,
-		'helpwizard'          => true
-	),
-	'sql'                     => "varchar(255) NOT NULL default ''",
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_aufsteiger'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_aufsteiger'],
+	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chessresults_round'],
+	'exclude'                 => true,
 	'inputType'               => 'text',
-	'eval'                    => array
-	(
-		'tl_class'            => 'long',
-		'maxlength'           => 255,
-		'helpwizard'          => true,
-	),
-	'explanation'             => 'chesstable_aufsteiger',
-	'sql'                     => "varchar(255) NOT NULL default ''",
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_absteiger'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_absteiger'],
-	'inputType'               => 'text',
-	'eval'                    => array
-	(
-		'tl_class'            => 'long',
-		'maxlength'           => 255,
-		'helpwizard'          => true,
-	),
-	'explanation'             => 'chesstable_absteiger',
-	'sql'                     => "varchar(255) NOT NULL default ''",
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_markieren'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_markieren'],
-	'inputType'               => 'text',
-	'eval'                    => array
-	(
-		'tl_class'            => 'long',
-		'maxlength'           =>255,
-		'helpwizard'          => true,
-	),
-	'explanation'             => 'chesstable_markieren',
-	'sql'                     => "varchar(255) NOT NULL default ''",
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_namendrehen'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_namendrehen'],
-	'inputType'               => 'checkbox',
 	'eval'                    => array
 	(
 		'tl_class'            => 'w50',
-		'isBoolean'           => true,
-		'helpwizard'          => true,
+		'maxlength'           => 3,
 	),
-	'explanation'             => 'Hilfetext',
-	'sql'                     => "char(1) NOT NULL default ''",
+	'sql'                     => "smallint(3) unsigned NOT NULL default '0'"
 );
 
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_flaggen'] = array
+$GLOBALS['TL_DCA']['tl_content']['fields']['chessresults_cache'] = array
 (
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_flaggen'],
-	'inputType'               => 'checkbox',
-	'eval'                    => array('tl_class' => 'w50','isBoolean' => true),
-	'sql'                     => "char(1) NOT NULL default ''",
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_date'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_date'],
-	'inputType'               => 'checkbox',
-	'eval'                    => array('tl_class' => 'w50','isBoolean' => true),
-	'sql'                     => "char(1) NOT NULL default ''",
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_ende'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_ende'],
-	'inputType'               => 'text',
-	'exclude'                 => false,
-	'search'                  => true,
-	'sorting'                 => true,
-	'flag'                    => 8,
-	'filter'                  => true,
-	'eval'                    => array
-	(
-		'rgxp'                => 'date',
-		'datepicker'          => true,
-		'tl_class'            => 'w50 wizard',
-		'doNotCopy'           => true
-	),
-	'sql'                     => "varchar(11) NOT NULL default ''"
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_note'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_note'],
-	'inputType'               => 'text',
-	'eval'                    => array
-	(
-		'tl_class'            => 'w50 long',
-		'maxlength'           => 255,
-	),
-	'sql'                     => "varchar(255) NOT NULL default ''",
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_lightbox'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_lightbox'],
+	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chessresults_cache'],
 	'exclude'                 => true,
-	'inputType'               => 'checkbox',
-	'eval'                    => array('tl_class'=>'w50 clr','isBoolean' => true,'submitOnChange'=>true),
-	'sql'                     => "char(1) NOT NULL default ''",
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_hinweis'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_hinweis'],
-	'eval'                    => array('tl_class'=>'long clr'),
-	'input_field_callback'    => array('tl_chesstable', 'jshinweis'),
-);
-
-$GLOBALS['TL_DCA']['tl_content']['fields']['chesstable_linktext'] = array
-(
-	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chesstable_linktext'],
 	'inputType'               => 'text',
-	'eval'                    => array('tl_class'=> 'w50 clr', 'maxlength'=>32),
-	'sql'                     => "varchar(32) NOT NULL default ''",
+	'default'                 => 300,
+	'eval'                    => array
+	(
+		'tl_class'            => 'w50',
+		'maxlength'           => 5,
+	),
+	'sql'                     => "int(5) unsigned NOT NULL default 0"
 );
 
-class tl_chesstable extends Backend
-{
+$GLOBALS['TL_DCA']['tl_content']['fields']['chessresults_colsView'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chessresults_colsView'],
+	'exclude'                 => true,
+	'filter'                  => true,
+	'inputType'               => 'checkboxWizard',
+	'options'                 => &$GLOBALS['TL_LANG']['tl_content']['chessresults_colsView_array'],
+	'default'                 => is_array($GLOBALS['TL_LANG']['tl_content']['chessresults_colsView_array']) ? array_keys($GLOBALS['TL_LANG']['tl_content']['chessresults_colsView_array']) : '',
+	'eval'                    => array
+	(
+		'multiple'            => true,
+		'tl_class'            => 'w50'
+	),
+	'sql'                     => 'blob NULL'
+);
 
-	public function jshinweis(DataContainer $dc)
-	{
-		return '<div class="tl_message clr">
-			<p class="tl_info">'.sprintf($GLOBALS['TL_LANG']['tl_content']['includeTemplate'], 'j_colorbox').'</p>
-			</div>';
-	}
-}
+$GLOBALS['TL_DCA']['tl_content']['fields']['chessresults_live'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chessresults_live'],
+	'exclude'                 => true,
+	'default'                 => 1,
+	'inputType'               => 'checkbox',
+	'eval'                    => array
+	(
+		'tl_class'            => 'w50'
+	),
+	'sql'                     => "char(1) NOT NULL default ''"
+);
 
-?>
+$GLOBALS['TL_DCA']['tl_content']['fields']['chessresults_cacheStorage'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chessresults_cacheStorage'],
+	'sql'                     => 'blob NULL'
+);
+
+$GLOBALS['TL_DCA']['tl_content']['fields']['chessresults_type'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_content']['chessresults_type'],
+	'exclude'                 => true,
+	'filter'                  => true,
+	'inputType'               => 'radio',
+	'default'                 => 1,
+	'options'                 => &$GLOBALS['TL_LANG']['tl_content']['chessresults_type_array'],
+	'eval'                    => array
+	(
+		'mandatory'           => true,
+		'tl_class'            => 'w50'
+	),
+	'sql'                     => "char(1) NOT NULL default ''"
+);
+