+++ src/Resources/contao/languages/de/tl_content.php
@@ -11,32 +11,47 @@
  * @copyright Frank Hoppe 2013
  */
 
-$GLOBALS['TL_LANG']['tl_content']['chesstable_csv']['0'] = "CSV-Daten";
-$GLOBALS['TL_LANG']['tl_content']['chesstable_csv']['1'] = "Hier tragen Sie die kommaseparierten Daten ein. Die Spalten müssen mit einem Semikolon getrennt sein.";
-$GLOBALS['TL_LANG']['tl_content']['chesstable_file']['0'] = "Oder alternative Datei einbinden";
-$GLOBALS['TL_LANG']['tl_content']['chesstable_file']['1'] = "Wählen Sie eine CSV-Datei aus";
-$GLOBALS['TL_LANG']['tl_content']['chesstable_aufsteiger']['0'] = 'Aufsteiger';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_aufsteiger']['1'] = 'Tragen Sie hier die Zeilennummern der Aufsteiger mit Komma und Bindestrich getrennt ein. Bsp.: 1,3-5,11. Orientieren Sie sich dabei an den Zeilennummern der Textarea!';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_absteiger']['0'] = 'Absteiger';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_absteiger']['1'] = 'Tragen Sie hier die Zeilennummern der Absteiger mit Komma und Bindestrich getrennt ein. Bsp.: 1,3-5,11. Orientieren Sie sich dabei an den Zeilennummern der Textarea!';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_markieren']['0'] = 'Sonstige Zeilen';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_markieren']['1'] = 'Tragen Sie hier die Zeilennummern der zu markierenden Zeilen mit Komma und Bindestrich getrennt ein. Bsp.: 1,3-5,11. Orientieren Sie sich dabei an den Zeilennummern der Textarea!';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_namendrehen']['0'] = "Spielernamen drehen";
-$GLOBALS['TL_LANG']['tl_content']['chesstable_namendrehen']['1'] = "Der Inhalt der Spielerspalte wird am Komma getrennt und gedreht.";
-$GLOBALS['TL_LANG']['tl_content']['chesstable_flaggen']['0'] = 'Flaggen anzeigen';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_flaggen']['1'] = 'Nationsflaggen anzeigen, wenn Länderspalte vorhanden ist.';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_date']['0'] = 'Aktualisierungsdatum anzeigen';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_date']['1'] = 'Datum der letzten Aktualisierung dieser Tabelle bzw. des Inhaltselements anzeigen.';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_ende']['0'] = 'Turnierendedatum';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_ende']['1'] = 'Zeigt das Turnierendedatum an, wenn gesetzt';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_note']['0'] = 'Hinweistext';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_note']['1'] = 'Hinweistext, der unterhalb der Tabelle angezeigt wird';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_lightbox']['0'] = 'Lightbox-Ansicht';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_lightbox']['1'] = 'Ansicht der Tabelle in einer Lightbox';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_linktext']['0'] = "Text für den Lightbox-Link";
-$GLOBALS['TL_LANG']['tl_content']['chesstable_linktext']['1'] = "Text für den Lightbox-Link";
+$GLOBALS['TL_LANG']['tl_content']['chessresults_legend'] = 'Einstellungen';
 
-$GLOBALS['TL_LANG']['tl_content']['chesstable_legend_csv'] = 'CSV-Daten';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_legend_aufab'] = 'Markierungen von Zeilen';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_legend_lightbox'] = 'Tabelle in einer Lightbox';
-$GLOBALS['TL_LANG']['tl_content']['chesstable_legend_optionen'] = 'Weitere Optionen';
+$GLOBALS['TL_LANG']['tl_content']['chessresults_id'] = array('Turnier-ID ChessResults', 'Turnier-ID ChessResults');
+$GLOBALS['TL_LANG']['tl_content']['chessresults_mode'] = array('Liste', 'Art der Liste');
+$GLOBALS['TL_LANG']['tl_content']['chessresults_round'] = array('Runde', 'Nummer der Runde');
+$GLOBALS['TL_LANG']['tl_content']['chessresults_cache'] = array('Cache', 'Cachezeit in Sekunden angeben, 0 = kein Cache');
+$GLOBALS['TL_LANG']['tl_content']['chessresults_cacheStorage'] = array('Cachespeicher', 'Im Cachespeicher werden die von ChessResults geladenen Daten mit Zeitstempel gespeichert.');
+$GLOBALS['TL_LANG']['tl_content']['chessresults_colsView'] = array('Spalten einblenden', 'Diese Spalten werden eingeblendet, wenn sie in der Ansicht vorhanden sind.');
+$GLOBALS['TL_LANG']['tl_content']['chessresults_live'] = array('Live-Abfrage', 'Wenn aktiviert, werden die Daten von ChessResults geladen. Cachezeiten werden berücksichtigt. Ist eine Offlinedarstellung nicht möglich, erfolgt trotzdem eine einmalige Abfrage bei ChessResults. Die Live-Abfrage ist ungefähr 500 mal langsamer. Bitte deshalb immer cachen, wenn kein Offline-Modus möglich ist!');
+$GLOBALS['TL_LANG']['tl_content']['chessresults_type'] = array('Turnierart', 'Turnierart');
+
+$GLOBALS['TL_LANG']['tl_content']['chessresults_mode_array'] = array
+(
+	'1'  => 'Teilnehmerliste, alphabetisch',
+	'2'  => 'Teilnehmerliste, nach TWZ',
+	'3'  => 'Rangliste, nach Runde ... (Bitte Runde angeben!)',
+	'10' => 'Fortschrittstabelle',
+	'20' => 'Paarungen/Ergebnisse, nach Runde ... (Bitte Runde angeben!)',
+);
+
+$GLOBALS['TL_LANG']['tl_content']['chessresults_type_array'] = array
+(
+	'1'  => 'Schweizer System',
+);
+
+$GLOBALS['TL_LANG']['tl_content']['chessresults_colsView_array'] = array
+(
+	'Nr.|Snr'    => 'Nr.',
+	'Rg.'        => 'Pl.',
+	'Br.'        => 'Br.',
+	'EloI'       => 'Elo',
+	'EloN'       => 'DWZ',
+	'Land'       => 'Land',
+	'Name'       => 'Name',
+	''           => 'Titel',
+	'Verein/Ort' => 'Verein',
+	'FideID'     => 'FIDE-ID',
+	'Ergebnis'   => 'Erg.',
+	'[Ergebnis]' => '(Ergebnis)',
+	'Pkt.'       => 'Pkt.',
+	'Wtg1'       => 'Wtg1',
+	'Wtg2'       => 'Wtg2',
+	'Wtg3'       => 'Wtg3',
+);