+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Tabellen von ChessResults einbinden
 
+## Version 1.0.0 (2020-07-31)
+
+* Add Teilnehmerliste, alphabetisch und nach Startrang
+* Add Rangliste normal
+* Add Ergebnisse
+* Add Live-Modus, Cache-Dauer und Offline-Cache in tl_content
+* Spalten optional ausblenden
+* Schweizer-System-Turniere fertiggestellt
+
 ## Version 0.0.2 (2020-07-29)
 
 * Fix composer.json Komma-Fehler